Pull gene autocomplete into separate component
